LayoutTests/http/tests/appcache/abort-cache-onprogress.html

<html manifest="resources/abort-cache-onprogress.manifest">
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText()
    layoutTestController.waitUntilDone();
}

function log(message) {
    document.getElementById("result").innerHTML += message + "\n";
}

function onprogress(event) {
    log("loading resource: " + event.loaded + " / " + event.total);
    if (event.loaded == 3) 
        applicationCache.abort();
}

function onnoupdate() {
    log("FAILURE");
    log("noupdate");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function oncached() {
    log("FAILURE");
    log("CACHED");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function onupdateready() {
    log("FAILURE");
    log("UPDATEREADY");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function onerror() {
    log("SUCCESS");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

applicationCache.addEventListener('noupdate', onnoupdate, false);
applicationCache.addEventListener('cached', oncached, false);
applicationCache.addEventListener('error', onerror, false);
applicationCache.addEventListener('updateready', onupdateready, false);
applicationCache.addEventListener('progress', onprogress, false);
</script>

<div>This tests that download process was aborted after progress event.</div>
<div id="result"></div>
</html>
